add full group page functionality
