+++ src/frontend/components/GroupList.jsx
@@ -1,14 +1,37 @@
 import React from "react";
-import GroupInfo from "./GroupInfo.jsx";
+import GroupCard from "./GroupCard.jsx";
 
 function GroupList({ groups, selectGroup, deleteGroup }) {
-  return (
-    <div className={"group-container"}>
-      {groups.map((group, index) => (
-        <GroupInfo key={index} group={group} selectGroup={selectGroup} deleteGroup={deleteGroup} />
-      ))}
-    </div>
-  );
+    const handleDelete = (e) => {
+
+    }
+
+    return (
+        <div className={"group-container"}>
+            {groups.map((group, index) => (
+                <div key={index}>
+                    <GroupCard group={group} selectGroup={selectGroup} deleteGroup={deleteGroup} />
+                    <button
+                        className="group-btn"
+                        type="submit"
+                        onClick={() => selectGroup(group._id)}
+                    >
+                        Go To Group Page
+                    </button>
+                    <button
+                        className={"delete-btn"}
+                        type={"submit"}
+                        onClick={(e) => {
+                            e.preventDefault();
+                            deleteGroup(group._id);
+                        }}
+                    >
+                        Delete Group
+                    </button>
+                </div>
+            ))}
+        </div>
+    );
 }
 
 export default GroupList;
